feat: add logging
