+++ src/football_pipeline/io.py
@@ -4,7 +4,7 @@ import urllib.request
 import pandas as pd
 import yaml
 
-from football_pipeline.logger import log_func
+from football_pipeline.logger import log_func, logger
 from football_pipeline.utils import sanitize_url
 
 
@@ -24,13 +24,15 @@ def read_yaml(path: str) -> dict:
             data = yaml.safe_load(file)
 
         if data is None:
-            print(f"Warning: {path} is empty or invalid YAML. Returning empty dict.")
+            logger.warning(
+                f"Warning: {path} is empty or invalid YAML. Returning empty dict."
+            )
             return {}
 
         return data
 
     except Exception as e:
-        print(f"Error loading YAML from {path}: {e}")
+        logger.error(f"Error loading YAML from {path}: {e}")
         return {}
 
 
@@ -40,7 +42,7 @@ def write_to_parquet(df: pd.DataFrame, path: str) -> bool:
     Takes a pandas dataframe, converts to parquet format and saves it in the specified path.
     """
     if not path.endswith(".parquet"):
-        print(
+        logger.error(
             f"{path} is an invalid file name. Please specify a .parquet file extension"
         )
         return False
@@ -49,7 +51,7 @@ def write_to_parquet(df: pd.DataFrame, path: str) -> bool:
         df.to_parquet(path)
         return True
     except Exception as e:
-        print(f"{path} is not a valid path: {e}")
+        logger.error(f"{path} is not a valid path: {e}")
         return False
 
 
@@ -76,7 +78,7 @@ def extract_data(url_path: str, keys: list = None) -> dict[str, pd.DataFrame]:
             data = json.load(url)
 
     except Exception as e:
-        print(f"{e} for {url_path}")
+        logger.error(f"{e} for {url_path}")
         return [{"err": e, "path": url_path}]
 
     if isinstance(data, dict):
@@ -85,7 +87,7 @@ def extract_data(url_path: str, keys: list = None) -> dict[str, pd.DataFrame]:
                 try:
                     dfs[k] = pd.DataFrame(v)
                 except Exception as e:
-                    print(f"{e} for {k}")
+                    logger.error(f"{e} for {k}")
                     fails.append({"err": e, "key": k})
 
     if isinstance(data, list):
@@ -93,7 +95,7 @@ def extract_data(url_path: str, keys: list = None) -> dict[str, pd.DataFrame]:
             sanitized_key = sanitize_url(url_path)
             dfs[sanitized_key] = pd.DataFrame(data)
         except Exception as e:
-            print(e)
+            logger.error(e)
             fails.append({"err": e, "key": sanitized_key})
 
     if fails:
