+++ src/football_pipeline/logger/__init__.py
@@ -1,5 +1,41 @@
+import functools
 import logging
 import logging.config
+from collections.abc import Callable
+
+
+def log_func(func: Callable) -> Callable:
+    """Decorator function that prints input/output of logger functions.
+
+    Args:
+        func: The function that the input/output will be logged for.
+
+    Returns:
+        Callable: Wrapper of the original function with the args and kwargs, then the function's result.
+    """
+
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        logger.info(
+            {
+                "func": func.__name__,
+                "args": args,
+                "kwargs": kwargs,
+            }
+        )
+
+        res = func(*args, **kwargs)
+
+        logger.info(
+            {
+                "func": func.__name__,
+                "res": res,
+            }
+        )
+        return res
+
+    return wrapper
+
 
 LOGGING_CONFIG = {  # constant, meaning it should not be changed after definition
     "version": 1,  # future-proofing in case of modifications to the logger class
